CDAP-3079 adding support for RecordWritable<StructuredRecord>
This change introduces support for StructuredRecord as the
type for a RecordWritable, similar to how RecordScannable
can have StructuredRecord as its type. Changed Table
so that it implements RecordWritable<StructuredRecord>,
making it possible to write to Tables through Hive queries.

Most of the work involves moving code from etl into cdap and
reusing it to deserialize json as a StructuredRecord, and to
convert a StructuredRecord into a Put.

Includes some cleanup and fixes around the ContextManager used in
explore to avoid reusing a closed context object.